Extract shared examples from InMemory::CarStore specs.
